Refactor logging utilities paths
