+++ src/events/guildAuditLogEntryCreate/handler.js
@@ -3,7 +3,7 @@
 */
 import { AuditLogEvent } from 'discord-api-types/v10';
 import { Events } from 'discord.js';
-import { logger } from '../../util/logger.js';
+import { logger } from '../../util/logging/logger.js';
 import {
   describeDiscordEntity,
   formatMetadataKey,
